De-flake two animated custom property WPTs
These two tests use a duration of 1000s and a delay of -500s
to effectively start the animation at 50%. That's fine, but we're
not necessarily guaranteed that the animation hasn't progressed
at the time we're making the test assertions. Therefore, make
the animations initially "paused".

Making an animation initially paused can sometimes void a test,
if relevant behavior can only be exercised by a *playing* animation.
However, that is not an issue for these tests, since they primarily
care about cascade interaction which isn't affected by playing/paused.

Fixed: 385790290, 385816384
Change-Id: I409101af17670316aba7b0faa9104fc14dd3e755